se actualizo el codigo

+++ dbmanagerall/codigo.py
@@ -15,8 +15,6 @@ URL_API=os.environ.get("URL_API")
 maximo_dias_acumular=int(os.environ.get("DIAS_ACUMULAR"))
 connlocal = None
 cursorlocal=None
-# listaUsuariosServidor=[]
-# listaUsuariosLocal=[]
 listaHuellasServidor=[]
 listahuellaslocal=[]
 listaempleadosseguricel=[]
@@ -24,7 +22,6 @@ total=0
 fechahoy=None
 fechaayer=None
 dias_acumulados=[]
-# etapa=0
 total_ping = 0
 nroCaptahuellasConHuella=0
 nroCaptahuellasSinHuella=0
@@ -32,12 +29,6 @@ captahuella_actual=0
 TIEMPO_PING=int(os.environ.get('TIEMPO_PING'))
 TIEMPO_CAMBIOS=int(os.environ.get('TIEMPO_CAMBIOS'))
 TIEMPO_LOG=int(os.environ.get('TIEMPO_LOG'))
-# consultarTodo=True
-# consultaUsuarios=False
-# consultaHorarios=False
-# consultaHuellas=False
-# consultaTags=False
-# subirLog=True
 BorrarPeticionesListas= True
 
 ######################################
@@ -226,6 +217,7 @@ while True:
                     if tablaCambiada == 'Usuarios':
                         try:
                             try:
+                                banderaUsuario=True
                                 cursorlocal.execute('SELECT cedula, nombre, telegram_id, internet, wifi, captahuella, rfid, facial FROM web_usuarios WHERE cedula=%s',(cedulaUsuario,))
                                 usuario_local= cursorlocal.fetchall()
 
@@ -251,6 +243,7 @@ while True:
                                                     nroCaptahuellasSinHuella=nroCaptahuellasSinHuella+1
                                                 except:
                                                     print(f"fallo al conectar con la esp8266 con la ip:{captahuella}")
+                                                    banderaUsuario=False
                                         if nroCaptahuellasSinHuella == captahuella_actual:
                                             cursorlocal.execute('DELETE FROM web_huellas WHERE id_suprema=%s', (id_suprema,))
                                             connlocal.commit()
@@ -259,7 +252,6 @@ while True:
                                         cursorlocal.execute('DELETE FROM web_usuarios WHERE cedula=%s', (cedulaUsuario,))
                                         cursorlocal.execute('DELETE FROM web_horariospermitidos WHERE cedula_id=%s', (cedulaUsuario,))
                                         connlocal.commit()
-                                    request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                                 elif not usuarioLocal and usuarioServidor: 
                                     for consultajson in request_json_usuario:
                                         cedula=consultajson['cedula']
@@ -273,7 +265,6 @@ while True:
                                     cursorlocal.execute('''INSERT INTO web_usuarios (cedula, nombre, telegram_id, internet, wifi, captahuella, rfid, facial)
                                     VALUES (%s, %s, %s, %s, %s, %s, %s, %s)''', (cedula, nombre, telegram_id, internet, wifi, captahuella, rfid, facial))
                                     connlocal.commit()
-                                    request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                                 elif usuarioLocal and usuarioServidor:
                                     for consultajson in request_json_usuario:
                                         cedula=consultajson['cedula']
@@ -285,14 +276,18 @@ while True:
                                         facial=consultajson['reconocimientoFacial']
                                     cursorlocal.execute("UPDATE web_usuarios SET telegram_id=%s, internet=%s, wifi=%s, captahuella=%s, rfid=%s, facial=%s WHERE cedula=%s", (telegram_id,internet,wifi,captahuella,rfid,facial,cedula))
                                     connlocal.commit()
-                                    request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                             except requests.exceptions.ConnectionError:
                                 print("fallo consultando api en usuarios")
+                                banderaUsuario=False
                         except Exception as e:
                             print(f"{e} - fallo total usuarios")
+                            banderaUsuario=False
+                        if banderaUsuario:
+                            requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                     elif tablaCambiada == 'Horarios':
                         try:
                             try:
+                                banderaHorario=True
                                 request_json_horarios = requests.get(url=f'{URL_API}obtenerhorariosindividualapi/{CONTRATO}/{cedulaUsuario}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3).json()
                                         
                                 horariosServidor=[]
@@ -323,11 +318,14 @@ while True:
                                         dia=horariosLocaliterar[3]
                                         cursorlocal.execute('DELETE FROM web_horariospermitidos WHERE entrada=%s AND salida=%s AND cedula_id=%s AND dia=%s',(entrada, salida, cedula, dia))
                                         connlocal.commit()
-                                request_json_horarios = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                             except requests.exceptions.ConnectionError:
-                                    print("fallo consultando api en horarios")
+                                print("fallo consultando api en horarios")
+                                banderaHorario=False
                         except Exception as e:
                             print(f"{e} - fallo total horarios")
+                            banderaHorario=False
+                        if banderaHorario:
+                            requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                     elif tablaCambiada == 'Huellas':
                         try:
                             try:
@@ -413,15 +411,18 @@ while True:
                                                     requests.get(url=f'{captahuella}/quitar/{id_suprema_hex}', timeout=3)
                                                 except:
                                                     print(f"fallo al conectar con la esp8266 con la ip:{captahuella}")
-                                if banderaHuella:
-                                    requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                             except requests.exceptions.ConnectionError:
-                                    print("fallo consultando api en huellas")
+                                print("fallo consultando api en huellas")
+                                banderaHuella=False
                         except Exception as e:
                             print(f"{e} - fallo total huellas")
+                            banderaHuella=False
+                        if banderaHuella:
+                            requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                     elif tablaCambiada == 'Tags':
                         try:
                             try:
+                                banderaTag=True
                                 cursorlocal.execute('SELECT epc, cedula FROM web_tagsrfid WHERE cedula=%s', (cedulaUsuario,))
                                 tags_local= cursorlocal.fetchall()
                                 
@@ -448,13 +449,16 @@ while True:
                                         cedula=taglocaliterar[1]
                                         cursorlocal.execute('DELETE FROM web_tagsrfid WHERE epc=%s AND cedula=%s',(epc, cedula))
                                         connlocal.commit()
-                                request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                             except requests.exceptions.ConnectionError:
-                                    print("fallo consultando api en tags")
+                                print("fallo consultando api en tags")
+                                banderaTag=False
                         except Exception as e:
                             print(f"{e} - fallo total tags")
+                            banderaTag=False
+                        if banderaTag:
+                            requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                     else:
-                        request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
+                        requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                 t1_cambios=tm.perf_counter()
             
             if total_log > TIEMPO_LOG:
@@ -499,9 +503,6 @@ while True:
                         if nro_int_local != nro_int_servidor:
 
                             for interaccion in interacciones_local:
-                                # try:
-                                #     listaLogsServidor.index(interaccion)
-                                # except ValueError:
                                 if not interaccion in listaLogsServidor:
                                     nombre=interaccion[0]
                                     fecha=interaccion[1]
@@ -548,7 +549,7 @@ while True:
                                             cursorlocal.execute('DELETE FROM solicitud_aperturas WHERE id=%s', (idapertura,))
                                             connlocal.commit()
                                     except requests.exceptions.ConnectionError:
-                                        print("fallo consultando api de periciones de aperturas")
+                                        print("fallo consultando api de peticiones de aperturas")
                                 elif not peticionDesdeInternet:# and feedbackPeticion:
                                     cursorlocal.execute('DELETE FROM solicitud_aperturas WHERE id=%s', (idapertura,))
                                     connlocal.commit()
